ActionとStoreの使用をViewModelに寄せる
