+++ Qiita-Training/View/Tabs/Article/ArticleList/ArticleListViewController.swift
@@ -13,25 +13,14 @@ import RxCocoa
 class ArticleListViewController: UIViewController {
 
     @IBOutlet private weak var tableView: UITableView!
-    private let dataSouce = ArticleListTableViewDataSouce()
-    private let disposeBag = DisposeBag()
+    private lazy var viewModel = ArticleListViewModel()
+    private lazy var dataSouce = ArticleListTableViewDataSouce(viewModel: viewModel)
     var qiitaAPIType: QiitaAPI?
     
     override func viewDidLoad() {
         super.viewDidLoad()
 
         dataSouce.configure(tableView: tableView)
-        observeArticleStore()
-    }
-    
-    private func observeArticleStore() {
-        ArticleStore.shared.article.asObservable()
-            .observeOn(MainScheduler.instance)
-            .filter { [weak self] in $0.1 == self?.qiitaAPIType }
-            .subscribe(onNext: { [weak self] article in
-                self?.dataSouce.articleList = article.0
-                self?.tableView.reloadData()
-            })
-            .disposed(by: disposeBag)
+        viewModel.observeArticleStore(qiitaAPIType: qiitaAPIType)
     }
 }
